# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Aaron Blog API 文檔'
  description: 'Aaron 個人博客系統的 API 文檔，提供文章、分類、標籤等功能的 RESTful API 介面。'
  contact:
    name: 'Aaron Blog 技術支援'
    email: admin@aaronblog.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080'
    description: 'Aaron Blog API 伺服器'
paths:
  /api/admin/articles:
    get:
      tags:
        - Admin
      summary: 獲取當前用戶的文章列表
      description: 獲取當前登入用戶的所有文章，包含草稿和已發布的文章，支援分頁、排序、搜尋功能
      operationId: getAdminArticles
      parameters:
        -
          name: page
          in: query
          description: 頁碼
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
            example: 1
        -
          name: per_page
          in: query
          description: 每頁筆數
          required: false
          schema:
            type: integer
            default: 10
            maximum: 50
            minimum: 1
            example: 10
        -
          name: sort_by
          in: query
          description: 排序欄位
          required: false
          schema:
            type: string
            default: created_at
            enum:
              - created_at
              - title
              - status
            example: created_at
        -
          name: sort_direction
          in: query
          description: 排序方向
          required: false
          schema:
            type: string
            default: desc
            enum:
              - asc
              - desc
            example: desc
        -
          name: search
          in: query
          description: 搜尋關鍵字（標題、內容）
          required: false
          schema:
            type: string
            maxLength: 100
            example: Laravel
        -
          name: status
          in: query
          description: 文章狀態篩選
          required: false
          schema:
            type: string
            enum:
              - draft
              - published
            example: published
      responses:
        200:
          description: 成功獲取文章列表
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: 用戶文章列表
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        title:
                          type: string
                          example: 'Laravel 最佳實踐'
                        slug:
                          type: string
                          example: laravel-best-practices
                        description:
                          type: string
                          example: '分享 Laravel 開發的最佳實踐'
                        status:
                          type: string
                          example: published
                        created_at:
                          type: string
                          format: date-time
                          example: '2024-03-20T12:00:00.000000Z'
                        updated_at:
                          type: string
                          format: date-time
                          example: '2024-03-20T14:30:00.000000Z'
                      type: object
                  pagination:
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 10
                      total:
                        type: integer
                        example: 25
                      last_page:
                        type: integer
                        example: 3
                    type: object
                type: object
        401:
          description: 未授權
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Unauthenticated
                type: object
        422:
          description: 驗證失敗
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 422
                  message:
                    type: string
                    example: 'The given data was invalid'
                type: object
      security:
        -
          sessionAuth: []
          csrfToken: []
    post:
      tags:
        - Admin
      summary: 創建新文章
      description: 創建新的文章，歸屬於當前登入用戶
      operationId: createAdminArticle
      requestBody:
        description: 文章資料
        required: true
        content:
          application/json:
            schema:
              required:
                - title
                - slug
                - description
                - content
                - status
              properties:
                title:
                  description: 文章標題（唯一）
                  type: string
                  maxLength: 255
                  example: 'Laravel 最佳實踐'
                slug:
                  description: 'URL slug（唯一，只能包含字母、數字、破折號和底線）'
                  type: string
                  maxLength: 255
                  example: laravel-best-practices
                description:
                  description: 文章描述
                  type: string
                  maxLength: 255
                  example: '分享 Laravel 開發的最佳實踐'
                content:
                  description: '文章內容（支援 Markdown）'
                  type: string
                  example: '# Laravel 最佳實踐\n\n這是文章內容...'
                category_id:
                  description: '分類 ID（可選）'
                  type: integer
                  example: 1
                status:
                  description: 文章狀態
                  type: string
                  enum:
                    - draft
                    - published
                  example: published
                tags:
                  description: '標籤 ID 陣列（可選）'
                  type: array
                  items:
                    type: integer
                    example: 1
              type: object
      responses:
        201:
          description: 文章創建成功
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: 操作成功
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      title:
                        type: string
                        example: 'Laravel 最佳實踐'
                      slug:
                        type: string
                        example: laravel-best-practices
                      description:
                        type: string
                        example: '分享 Laravel 開發的最佳實踐'
                      content:
                        type: string
                        example: '# Laravel 最佳實踐...'
                      status:
                        type: string
                        example: published
                      created_at:
                        type: string
                        format: date-time
                        example: '2024-03-20T12:00:00.000000Z'
                      updated_at:
                        type: string
                        format: date-time
                        example: '2024-03-20T12:00:00.000000Z'
                    type: object
                type: object
        401:
          description: 未授權
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Unauthenticated
                type: object
        422:
          description: 驗證失敗
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 422
                  message:
                    type: string
                    example: 'The given data was invalid'
                  errors:
                    properties:
                      title:
                        type: array
                        items:
                          type: string
                          example: 標題已被使用
                      slug:
                        type: array
                        items:
                          type: string
                          example: 'slug 已被使用'
                    type: object
                type: object
      security:
        -
          sessionAuth: []
          csrfToken: []
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      description: '使用 Laravel Session Cookie 進行身份驗證'
      name: laravel_session
      in: cookie
    csrfToken:
      type: apiKey
      description: 'CSRF 防護 Token，用於 POST/PUT/DELETE 等寫入操作'
      name: X-XSRF-TOKEN
      in: header
...